chore(changelog): update to version 1.0.3 with bug fixes and demo improvements
