+++ src/demo/Demo.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { defineAsyncComponent } from 'vue'
 import { TvDemo } from '@todovue/tv-demo'
-import { demos } from "../utils/mocks.js";
+import { demos } from './utils/mocks.js'
 
 const TvSearch = defineAsyncComponent(() => import('../components/TvSearch.vue'))
 </script>
@@ -14,7 +14,7 @@ const TvSearch = defineAsyncComponent(() => import('../components/TvSearch.vue')
     npmInstall="@todovue/tv-search"
     sourceLink="https://github.com/TODOvue/tv-search"
     urlClone="https://github.com/TODOvue/tv-search.git"
-    version="1.0.2"
+    version="1.0.3"
   ></tv-demo>
 </template>
 <style scoped>
