+++ src/demo/utils/demos/default.vue
@@ -1,10 +1,10 @@
 <template>
   <!-- You can open the search component with Ctrl + k or Command + k -->
   <tv-search
-    placeholder="Type 'Blog'..."
+    placeholder="Type 'How'..."
     titleButton="Search"
     :results="results"
-    :customStyles="customStyles"
+    @search="console.log('Searching for:', $event)"
   />
 </template>
 
@@ -12,41 +12,34 @@
 import { ref } from 'vue';
 import { TvSearch } from '@todovue/tv-search';
 
-const customStyles = ref({
-  bgBody: "#0A4539",
-  bgInput: "#284780",
-  bgButton: "#80286E",
-  colorButton: "#D5B7B7",
-});
-
 const results = ref([
   {
     id: 1,
-    title: 'Blog - How to use Vue 3',
+    title: 'How to use Vue 3',
     description: 'Vue 3 is the latest version of Vue.js',
     url: 'https://todovue.com/blog/how-to-use-vue-3',
   },
   {
     id: 2,
-    title: 'Blog - How to use Vite',
+    title: 'How to use Vite',
     description: 'Vite is a build tool for modern web development',
     url: 'https://todovue.com/blog/how-to-use-vite',
   },
   {
     id: 3,
-    title: 'Blog - How to use Pinia',
+    title: 'How to use Pinia',
     description: 'Pinia is a modern store for Vue 3',
     url: 'https://todovue.com/blog/how-to-use-pinia',
   },
   {
     id: 4,
-    title: 'Blog - How to use Windi CSS',
+    title: 'How to use Windi CSS',
     description: 'Windi CSS is a utility-first CSS framework',
     url: 'https://todovue.com/blog/how-to-use-windi-css',
   },
   {
     id: 5,
-    title: 'Blog - How to use Vitesse',
+    title: 'How to use Vitesse',
     description: 'Vitesse is a Vue 3 starter template',
     url: 'https://todovue.com/blog/how-to-use-vitesse',
   }
